backend/requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-dotenv
PyJWT
supabase
alembic
sqladmin
itsdangerous
